% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty.hall.3doors.R
\name{change_door}
\alias{change_door}
\title{Step 4: Change Doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{No}{arguments are used}
}
\value{
The door that the contestant chooses, either the same door or a new door
}
\description{
The contestant chooses to either switch their door option or keep their original door
}
\details{
The contestant will then have a 66\% of choosing a door with a car behind it.
}
\examples{
change door 
}
