% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty.hall.3doors.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine If the Contestant Has Won}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{No}{arguments are used}
}
\value{
If the contestant has won
}
\description{
By analyzing the contestant's choices we will see if they chose the car or the goat
}
\details{
We will see if the contestant chose the door with the car behind it or the goat door by seeing if they switched or kept their original door choice
}
\examples{
Determine winner 
}
