% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty.hall.3doors.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Step 3: Host Opens Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{No}{arguments are used}
}
\value{
a door that has a goat behind it
}
\description{
The host will open a door with a goat behind it.
}
\details{
The host will always choose a door with a goat behind it but it cannot be the door that the contestant already chose.
}
\examples{
open a goat door 
}
