% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty.hall.3doors.R
\name{play_game}
\alias{play_game}
\title{test the game}
\usage{
play_game()
}
\arguments{
\item{no}{arguments used}
}
\value{
overall outcome of game
}
\description{
play the game
}
\details{
Test the game from start to finish to see if the contestant has won
}
\examples{
game outcome
}
