% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty.hall.3doors.R
\name{play_n_games}
\alias{play_n_games}
\title{Document results by testing game}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{No}{arguments used}
}
\value{
Table of results
}
\description{
Document number of times the contestant wins or loses
}
\details{
Create a table in order to save the outcome of contestant's choices and seeing how many times they win versus lose
}
\examples{
results in a table  
}
