% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty.hall.3doors.R
\name{select_door}
\alias{select_door}
\title{Step 2: Contestant Selects a Door}
\usage{
select_door()
}
\arguments{
\item{no}{arguments are used}
}
\value{
The function returns one random door chosen by the contestant
}
\description{
The contestant chooses their first door
}
\details{
Need to write a function to select a random door
}
\examples{
select a door 
}
